pyproject.toml
[project]
name = "mcp-workshop"
version = "0.1.0"
description = "MCP workshop"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Yingting Huang" }]
keywords = ["mcp", "llm", "azure", "apim", "stdio", "http", "streamable"]
license = { text = "MIT" }
dependencies = [
    "aiohttp>=3.12.15",
    "mcp[cli]>=1.13.0",
    "msal>=1.33.0",
    "httpx>=0.27.2",
    "openai>=1.100.2",
    "pyjwt[crypto]>=2.10.1",
    "requests>=2.32.5",
    "yfinance>=0.2.65",
    "azure-ai-projects==1.1.0b3",
    "azure-identity>=1.24.0",
    "azure-ai-agents==1.2.0b3",
    "azure-functions>=1.23.0",
    "newsapi-python>=0.2.7",
    "ipywidgets>=8.1.7",
    "ipykernel>=6.30.1",
    "smolagents[mcp,toolkit]>=1.21.2",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
]
